Stage 2 finalization. Few modules from ppd_sync grouped. Now there is the tests separately for main function, pd item processing, and by module. Conftest now have reference item, not simulated, which is important. (2)
